+++ src/main/java/com/chirkov/restApiRestaurantBussines/repositories/StateFromTablesRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface StateFromTablesRepository extends JpaRepository<StateFromTable,Integer> {
+public interface StateFromTablesRepository extends JpaRepository<StateFromTable,Long> {
     Optional<StateFromTable> findStateFromTablesByName(String tableName);
     Optional<StateFromTable> findStateFromTablesByValue(StateFromTableEnum value);
 }
